02- Pinecone-Image/requirements.txt
pinecone-client==2.2.
openai
fastapi==0.95.1
uvicorn==0.20.0
sentence_transformers==2.2.2
pandas==2.0.3
python-dotenv
timm
tqdm
requests
python-multipart